refactor: reuse shared GlassCardModifier across settings and cities
