+++ SkyCast — Weather App/Weather-Helper/SettingsView.swift
@@ -247,40 +247,7 @@ private struct SettingRowData {
     let systemImage: String
 }
 
-private struct SettingsGlassCardModifier: ViewModifier {
-    @Environment(\.colorScheme) private var colorScheme
-    let cornerRadius: CGFloat
-    let extraDarkTint: Double
-
-    func body(content: Content) -> some View {
-        let shape = RoundedRectangle(cornerRadius: cornerRadius, style: .continuous)
-
-        content
-            .background {
-                ZStack {
-                    shape
-                        .fill(.ultraThinMaterial)
-
-                    shape
-                        .fill(colorScheme == .dark ? Color.black.opacity(0.34) : Color.white.opacity(0.10))
-
-                    shape
-                        .fill(Color.black.opacity(extraDarkTint))
-                }
-            }
-            .shadow(color: .black.opacity(colorScheme == .dark ? 0.24 : 0.07), radius: 14, x: 0, y: 8)
-            .overlay {
-                shape
-                    .stroke(.white.opacity(colorScheme == .dark ? 0.10 : 0.28), lineWidth: 1)
-            }
-    }
-}
-
-private extension View {
-    func glassCard(cornerRadius: CGFloat = 22, extraDarkTint: Double = 0) -> some View {
-        modifier(SettingsGlassCardModifier(cornerRadius: cornerRadius, extraDarkTint: extraDarkTint))
-    }
-}
+// MARK: - Preview
 
 #Preview {
     SettingsView(viewModel: WeatherViewModel())
